+++ dio/lib/src/options.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:meta/meta.dart';
 
 import 'adapter.dart';
@@ -76,14 +78,14 @@ enum ListFormat {
 typedef ValidateStatus = bool Function(int? status);
 
 /// The type of a response decoding callback.
-typedef ResponseDecoder = String? Function(
+typedef ResponseDecoder = FutureOr<String?> Function(
   List<int> responseBytes,
   RequestOptions options,
   ResponseBody responseBody,
 );
 
-/// The type of a response encoding callback.
-typedef RequestEncoder = List<int> Function(
+/// The type of a request encoding callback.
+typedef RequestEncoder = FutureOr<List<int>> Function(
   String request,
   RequestOptions options,
 );
@@ -98,14 +100,14 @@ mixin OptionsMixin {
   /// List values use the default [ListFormat.multiCompatible].
   /// The value can be overridden per parameter by adding a [ListParam]
   /// object wrapping the actual List value and the desired format.
-  late Map<String, dynamic> queryParameters;
+  late Map<String, dynamic /*String|Iterable<String>*/ > queryParameters;
 
-  /// Timeout when opening url.
+  /// Timeout when opening a request.
   ///
-  /// [Dio] will throw the [DioException] with
+  /// Throws the [DioException] with
   /// [DioExceptionType.connectionTimeout] type when time out.
   ///
-  /// `null` meanings no timeout limit.
+  /// `null` or `Duration.zero` means no timeout limit.
   Duration? get connectTimeout => _connectTimeout;
   Duration? _connectTimeout;
 
@@ -128,6 +130,7 @@ class BaseOptions extends _RequestConfig with OptionsMixin {
     Map<String, dynamic>? queryParameters,
     Map<String, dynamic>? extra,
     Map<String, dynamic>? headers,
+    bool preserveHeaderCase = false,
     ResponseType? responseType = ResponseType.json,
     String? contentType,
     ValidateStatus? validateStatus,
@@ -146,6 +149,7 @@ class BaseOptions extends _RequestConfig with OptionsMixin {
           sendTimeout: sendTimeout,
           extra: extra,
           headers: headers,
+          preserveHeaderCase: preserveHeaderCase,
           responseType: responseType,
           contentType: contentType,
           validateStatus: validateStatus,
@@ -173,6 +177,7 @@ class BaseOptions extends _RequestConfig with OptionsMixin {
     Duration? sendTimeout,
     Map<String, Object?>? extra,
     Map<String, Object?>? headers,
+    bool? preserveHeaderCase,
     ResponseType? responseType,
     String? contentType,
     ValidateStatus? validateStatus,
@@ -193,6 +198,7 @@ class BaseOptions extends _RequestConfig with OptionsMixin {
       sendTimeout: sendTimeout ?? this.sendTimeout,
       extra: extra ?? Map.from(this.extra),
       headers: headers ?? Map.from(this.headers),
+      preserveHeaderCase: preserveHeaderCase ?? this.preserveHeaderCase,
       responseType: responseType ?? this.responseType,
       contentType: contentType ?? this.contentType,
       validateStatus: validateStatus ?? this.validateStatus,
@@ -216,6 +222,7 @@ class Options {
     Duration? receiveTimeout,
     this.extra,
     this.headers,
+    this.preserveHeaderCase,
     this.responseType,
     this.contentType,
     this.validateStatus,
@@ -238,6 +245,7 @@ class Options {
     Duration? receiveTimeout,
     Map<String, Object?>? extra,
     Map<String, Object?>? headers,
+    bool? preserveHeaderCase,
     ResponseType? responseType,
     String? contentType,
     ValidateStatus? validateStatus,
@@ -274,6 +282,7 @@ class Options {
       receiveTimeout: receiveTimeout ?? this.receiveTimeout,
       extra: extra ?? effectiveExtra,
       headers: headers ?? effectiveHeaders,
+      preserveHeaderCase: preserveHeaderCase ?? this.preserveHeaderCase,
       responseType: responseType ?? this.responseType,
       contentType: contentType ?? this.contentType,
       validateStatus: validateStatus ?? this.validateStatus,
@@ -322,6 +331,7 @@ class Options {
       baseUrl: baseOpt.baseUrl,
       path: path,
       data: data,
+      preserveHeaderCase: preserveHeaderCase ?? baseOpt.preserveHeaderCase,
       sourceStackTrace: sourceStackTrace ?? StackTrace.current,
       connectTimeout: baseOpt.connectTimeout,
       sendTimeout: sendTimeout ?? baseOpt.sendTimeout,
@@ -353,15 +363,24 @@ class Options {
   /// HTTP request headers.
   ///
   /// The keys of the header are case-insensitive,
-  /// e.g.: content-type and Content-Type will be treated as the same key.
+  /// e.g.: `content-type` and `Content-Type` will be treated as the same key.
   Map<String, dynamic>? headers;
+
+  /// Whether the case of header keys should be preserved.
+  ///
+  /// Defaults to false.
+  ///
+  /// This option WILL NOT take effect on these circumstances:
+  /// - XHR ([HttpRequest]) does not support handling this explicitly.
+  /// - The HTTP/2 standard only supports lowercase header keys.
+  bool? preserveHeaderCase;
 
   /// Timeout when sending data.
   ///
-  /// [Dio] will throw the [DioException] with
+  /// Throws the [DioException] with
   /// [DioExceptionType.sendTimeout] type when timed out.
   ///
-  /// `null` meanings no timeout limit.
+  /// `null` or `Duration.zero` means no timeout limit.
   Duration? get sendTimeout => _sendTimeout;
   Duration? _sendTimeout;
 
@@ -374,13 +393,16 @@ class Options {
 
   /// Timeout when receiving data.
   ///
-  /// The timeout represents the timeout during data transfer of each bytes,
-  /// rather than the overall timing during the receiving.
+  /// The timeout represents:
+  ///  - a timeout before the connection is established
+  ///    and the first received response bytes.
+  ///  - the duration during data transfer of each byte event,
+  ///    rather than the total duration of the receiving.
   ///
-  /// [Dio] will throw the [DioException] with
-  /// [DioExceptionType.receiveTimeout] type when time out.
+  /// Throws the [DioException] with
+  /// [DioExceptionType.receiveTimeout] type when timed out.
   ///
-  /// `null` meanings no timeout limit.
+  /// `null` or `Duration.zero` means no timeout limit.
   Duration? get receiveTimeout => _receiveTimeout;
   Duration? _receiveTimeout;
 
@@ -396,20 +418,20 @@ class Options {
   /// {@macro dio.interceptors.ImplyContentTypeInterceptor}
   String? contentType;
 
-  /// Indicates the type of data that the server will respond with options.
+  /// The type of data that [Dio] handles with options.
   ///
-  /// The default value is [ResponseType.json], [Dio] will parse response string
-  /// to JSON object automatically when the content-type of response is
-  /// [Headers.jsonContentType].
+  /// The default value is [ResponseType.json].
+  /// [Dio] will parse response string to JSON object automatically
+  /// when the content-type of response is [Headers.jsonContentType].
   ///
-  /// If you want to receive response data with binary bytes, use `stream`.
-  ///
-  /// If you want to receive the response data with String, use `plain`.
-  ///
-  /// If you want to receive the response data with original bytes, use `bytes`.
+  /// See also:
+  ///  - `plain` if you want to receive the data as `String`.
+  ///  - `bytes` if you want to receive the data as the complete bytes.
+  ///  - `stream` if you want to receive the data as streamed binary bytes.
   ResponseType? responseType;
 
-  /// Defines whether the request is succeed with the given status code.
+  /// Defines whether the request is considered to be successful
+  /// with the given status code.
   /// The request will be treated as succeed if the callback returns true.
   ValidateStatus? validateStatus;
 
@@ -440,12 +462,14 @@ class Options {
   /// Defaults to true.
   bool? persistentConnection;
 
-  /// The default request encoder is utf8encoder, you can set custom
-  /// encoder by this option.
+  /// The type of a request encoding callback.
+  ///
+  /// Defaults to [Utf8Encoder].
   RequestEncoder? requestEncoder;
 
-  /// The default response decoder is utf8decoder, you can set custom
-  /// decoder by this option, it will be used in [Transformer].
+  /// The type of a response decoding callback.
+  ///
+  /// Defaults to [Utf8Decoder].
   ResponseDecoder? responseDecoder;
 
   /// Indicates the format of collection data in request query parameters and
@@ -472,6 +496,7 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
     String? baseUrl,
     Map<String, dynamic>? extra,
     Map<String, dynamic>? headers,
+    bool? preserveHeaderCase,
     ResponseType? responseType,
     String? contentType,
     ValidateStatus? validateStatus,
@@ -491,6 +516,7 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
           receiveTimeout: receiveTimeout,
           extra: extra,
           headers: headers,
+          preserveHeaderCase: preserveHeaderCase,
           responseType: responseType,
           contentType: contentType,
           validateStatus: validateStatus,
@@ -523,6 +549,7 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
     CancelToken? cancelToken,
     Map<String, dynamic>? extra,
     Map<String, dynamic>? headers,
+    bool? preserveHeaderCase,
     ResponseType? responseType,
     String? contentType,
     ValidateStatus? validateStatus,
@@ -559,6 +586,7 @@ class RequestOptions extends _RequestConfig with OptionsMixin {
       cancelToken: cancelToken ?? this.cancelToken,
       extra: extra ?? Map.from(this.extra),
       headers: headers ?? Map.from(this.headers),
+      preserveHeaderCase: preserveHeaderCase ?? this.preserveHeaderCase,
       responseType: responseType ?? this.responseType,
       validateStatus: validateStatus ?? this.validateStatus,
       receiveDataWhenStatusError:
@@ -636,6 +664,7 @@ class _RequestConfig {
     String? method,
     Map<String, dynamic>? extra,
     Map<String, dynamic>? headers,
+    bool? preserveHeaderCase,
     String? contentType,
     ListFormat? listFormat,
     bool? followRedirects,
@@ -651,6 +680,7 @@ class _RequestConfig {
         assert(sendTimeout == null || !sendTimeout.isNegative),
         _sendTimeout = sendTimeout,
         method = method ?? 'GET',
+        preserveHeaderCase = preserveHeaderCase ?? false,
         listFormat = listFormat ?? ListFormat.multi,
         extra = extra ?? {},
         followRedirects = followRedirects ?? true,
@@ -689,6 +719,8 @@ class _RequestConfig {
       _headers[Headers.contentTypeHeader] = _defaultContentType;
     }
   }
+
+  late bool preserveHeaderCase;
 
   Duration? get sendTimeout => _sendTimeout;
   Duration? _sendTimeout;